requirements.txt
Flask==3.0.0
line-bot-sdk>=3.13.0
google-generativeai==0.3.2
gunicorn==21.2.0
icalendar==5.0.11
python-dateutil==2.8.2
pytz==2023.3
